requirements.txt
flask==3.0.1
pymongo==4.7.1
requests==2.31.0
python-dotenv==1.0.1